requirements.txt
pyyaml
numpy
pandas
matplotlib
scipy
torch
torchvision
scikit-learn
pillow
mlflow
